watermark-service/requirements.txt
# FashionGuard Watermark Service Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Scientific Computing
scikit-image==0.22.0
scipy==1.12.0

# Cryptography
cryptography==42.0.0

# Error Correction
reedsolo==1.7.0

# Settings Management
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
